my-app.python/pyproject.toml
[tool.poetry]
name = "my-app.python"
version = "0.1.0"
description = ""
authors = ["shin5ok <shin5ok@55mp.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.6"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
google-cloud-firestore = "^2.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
